Create privoxyprotection.sh
This script run as a cron will search for the word "porn" in logs for Privoxy server. When found it will take the last IP address and add it to denied-access list for privoxy. This is the most basic variation of this that I have. You can do amazing things with this. Make your own algorithm and find people up to no good.